+++ philo/utils.c
@@ -41,7 +41,7 @@ int	positive_atoi(const char *str)
 }
 
 /* Get the current time in milliseconds since 1970-01-01 00:00:00*/
-long	get_time(void)
+int	get_time(void)
 {
 	struct timeval	time_value;
